[GameServer] 64비트 필드 ID 체계 구축 및 인스턴스 자동 확장 시스템 구현
- 64비트 필드 ID 체계를 확립하여 상위 16비트 MapID와 하위 48비트 InstanceID로 구성함
- FieldManager에 이중 인덱싱 구조(_fieldIdInstanceMap, _fields)를 도입하여 고유 ID 기반 검색 및 맵별 버킷 관리 최적화함
- GetField(mapId) 요청 시 플레이어 수용 인원(MAX_PLAYERS)을 체크하여 자동 인스턴스 생성 및 부하 분산 로직 구현함
- 필드 내 플레이어 부재 시 10초 뒤 인스턴스를 자동 파괴하는 가비지 컬렉션 로직을 Token 기반으로 안전하게 구축함
- Field 클래스의 GetMapID, GetInstanceID 내 비트 연산 및 마스킹 오류를 수정하여 ID 분석 정확도 확보함
- 오늘 작업 요약 문서([2026-03-12]_Summary.md)를 프로젝트에 추가함

+++ Server/GameServer/Contents/Field/FieldData.cpp
@@ -8,7 +8,7 @@ namespace fs = std::filesystem;
 
 FieldData::FieldData(nlohmann::json j)
 {
-    _fieldId = j["FieldId"].get<int32>();
+    _fieldId = j["FieldId"].get<uint16>();
     _mapName = j["MapName"].get<std::string>();
 
     if(j.contains("PlayerStart"))
